+++ src/Sql/Sql.Test/ScenarioTests/TransparentDataEncryptionCrudTests.ps1
@@ -75,7 +75,7 @@ function Test-GetTransparentDataEncryption
 			-DatabaseName $db.DatabaseName -State Disabled
 		Assert-AreEqual $tde3.State Disabled
 
-		Start-Sleep -s 1
+		Start-TestSleep -Seconds 1
 
 		$tdeActivity = Get-AzSqlDatabaseTransparentDataEncryptionActivity -ResourceGroupName $server.ResourceGroupname `
 			-ServerName $server.ServerName -DatabaseName $db.DatabaseName