+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Removed unused reference to resource identifier in role assignment create cmdlet
 
 ## Version 6.2.0
 * Fixed bug `-Password` overwrite `-PasswordProfile` in `New-AzADUser` [#19265]